Changed number type from tiny to small

+++ database/migrations/2023_09_25_131621_create_addresses_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->string('toponymic');
             $table->string('street_name');
-            $table->tinyInteger('number');
+            $table->smallint('number');
             $table->char('zip_code', 5);
             $table->string('city');
             $table->string('latitude');
